fix(darwin/system): use correct key repeat timings
Oh boy, that was hilarious. Was looking at the floating point values
used in the accessibility settings and those were less than 1, so I
assumed a setting of 1 made sense for these. I basically logged in to a
session where I couldn't type much of anything.

Found the right settings after manually adjusting the values in the
Settings app and just copied over the values for those.

lol