feat: split off darwin configs for okra host
Actually aggregating a bunch of small, relevant changes that were
required to support a second darwin host.

The major change is restructuring how modules of a given type are
imported. Now at the flake, a helper module that imports all role
modules of a given type is imported, and then an inline module is added
that sets module options to enable and configure those modules.

The other big thing is having current darwin configurations moved under
the danh user. After splitting off a `common` module for both `okra` and
`keylime` it became pretty apparent that as they were, host modules only
ever made sense in the context of a single users. If I were configuring
something like MacMini CI hosts or a media server, there might be a
better reason to have darwin modules for hosts. But for now, the
personal laptop configurations are gonna live as part of the user roles.

Included here is also a dependency update. It wasn't a mindful decision
-- I was flailing while trying to fix some segfault that came up as part
of an infinite import loop I caused. I figure I might as well keep it
since I already worked to figure out what was broken and how.